move clients/utils into utils
